backend/repl/pyproject.toml
[project]
name = "repl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "appnope==0.1.4",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asttokens==2.4.1",
    "async-lru==2.0.4",
    "async-timeout==4.0.3",
    "attrs==23.2.0",
    "babel==2.15.0",
    "beautifulsoup4==4.12.3",
    "bleach==6.1.0",
    "certifi==2024.6.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cmdstanpy==1.2.4",
    "comm==0.2.2",
    "contourpy==1.2.1",
    "cycler==0.12.1",
    "cython==3.0.11",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.2",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "dill>=0.4.0",
    "dnspython==2.6.1",
    "domdf-python-tools==3.9.0",
    "email-validator==2.2.0",
    "et-xmlfile==1.1.0",
    "exceptiongroup==1.2.1",
    "executing==2.0.1",
    "faker==33.1.0",
    "fastapi==0.111.0",
    "fastapi-cli==0.0.4",
    "fastjsonschema==2.20.0",
    "fonttools==4.53.0",
    "fqdn==1.5.1",
    "frozenlist==1.4.1",
    "gigachat==0.1.35",
    "h11==0.14.0",
    "hdbscan>=0.8.40",
    "holidays==0.57",
    "html5lib==1.1",
    "httpcore==1.0.5",
    "httptools==0.6.1",
    "httpx==0.27.0",
    "idna==3.7",
    "importlib-metadata==8.0.0",
    "importlib-resources==6.4.0",
    "ipykernel==6.29.5",
    "ipython==8.18.1",
    "ipywidgets==8.1.3",
    "isoduration==20.11.0",
    "jedi==0.19.1",
    "jinja2==3.1.4",
    "joblib==1.4.2",
    "json5==0.9.25",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.22.0",
    "jsonschema-specifications==2023.12.1",
    "jupyter==1.0.0",
    "jupyter-client==8.6.2",
    "jupyter-console==6.6.3",
    "jupyter-core==5.7.2",
    "jupyter-events==0.10.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.14.1",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.2.3",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.2",
    "jupyterlab-widgets==3.0.11",
    "kaleido==1.0.0",
    "kiwisolver==1.4.5",
    "langchain-gigachat>=0.3.0",
    "lxml==5.3.0",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "marshmallow==3.22.0",
    "mathematical==0.5.1.post1",
    "matplotlib==3.9.0",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.0.2",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "natsort==8.4.0",
    "nbclient==0.10.0",
    "nbconvert==7.16.4",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.2.1",
    "nltk==3.9.1",
    "notebook==7.2.1",
    "notebook-shim==0.2.4",
    "numpy==1.26.4",
    "openai==0.28.0",
    "openpyxl==3.1.5",
    "orjson==3.10.15",
    "overrides==7.7.0",
    "packaging==24.1",
    "pandas==2.2.2",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "patsy==0.5.6",
    "pexpect==4.9.0",
    "pillow==10.4.0",
    "platformdirs==4.2.2",
    "plotly==6.1.0",
    "pmdarima==2.0.4",
    "prometheus-client==0.20.0",
    "prompt-toolkit==3.0.47",
    "propcache==0.2.0",
    "prophet==1.1.5",
    "psutil==6.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.2",
    "pycparser==2.22",
    "pydantic==2.8.0",
    "pydantic-core==2.20.0",
    "pygments==2.18.0",
    "pyparsing==3.1.2",
    "python-dateutil==2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv==1.0.1",
    "python-json-logger==2.0.7",
    "python-multipart==0.0.9",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "pyzmq==26.0.3",
    "qtconsole==5.5.2",
    "qtpy==2.4.1",
    "referencing==0.35.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rich==13.7.1",
    "rpds-py==0.18.1",
    "scikit-learn==1.6.1",
    "scipy==1.13.1",
    "seaborn==0.13.2",
    "send2trash==1.8.3",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.5",
    "sqlalchemy==2.0.36",
    "stack-data==0.6.3",
    "stanio==0.5.1",
    "starlette==0.37.2",
    "statsmodels==0.14.3",
    "tabulate==0.9.0",
    "tenacity==8.4.2",
    "terminado==0.18.1",
    "threadpoolctl==3.5.0",
    "tinycss2==1.3.0",
    "tomli==2.0.1",
    "tornado==6.4.1",
    "tqdm==4.66.5",
    "traitlets==5.14.3",
    "typer==0.12.3",
    "types-python-dateutil==2.9.0.20240316",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "tzdata==2024.1",
    "ujson==5.10.0",
    "umap-learn>=0.5.9.post2",
    "uri-template==1.3.0",
    "urllib3==2.2.2",
    "uvicorn==0.29.0",
    "uvloop==0.19.0",
    "watchfiles==0.22.0",
    "wcwidth==0.2.13",
    "webcolors==24.6.0",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets==12.0",
    "widgetsnbextension==4.0.11",
    "wordcloud==1.9.4",
    "yarl==1.15.3",
    "zipp==3.19.2",
]
